requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
appscript==1.3.0
attrs==25.3.0
black==25.1.0
Brotli==1.1.0
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cobble==0.1.4
cryptography==44.0.2
cssselect2==0.8.0
deprecation==2.1.0
distro==1.9.0
docx2pdf==0.1.8
fastapi==0.115.11
flake8==7.1.2
fonttools==4.56.0
frozenlist==1.5.0
git-filter-repo==2.47.0
gitdb==4.0.12
gitdb2==4.0.2
GitPython==3.0.6
gotrue==1.3.1
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
Jinja2==3.1.6
jiter==0.8.2
lxml==5.3.1
mammoth==1.9.0
MarkupSafe==3.0.2
mccabe==0.7.0
multidict==6.1.0
mypy-extensions==1.0.0
oauthlib==3.2.2
openai==1.64.0
packaging==24.2
pathspec==0.12.1
pdfminer.six==20231228
pdfplumber==0.11.5
pdfrw==0.4
pillow==11.1.0
platformdirs==4.3.6
postgrest==0.11.0
propcache==0.3.0
pyan3==1.2.0
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pydyf==0.11.0
pyflakes==3.2.0
PyJWT==2.10.1
PyMuPDF==1.25.3
pypdfium2==4.30.1
pyphen==0.17.2
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==0.8.11
python-dotenv==1.0.1
python-multipart==0.0.20
python-telegram-bot==22.0
realtime==1.0.6
reportlab==4.3.1
requests==2.32.3
requests-oauthlib==2.0.0
ruff==0.11.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.39
starlette==0.46.1
storage3==0.6.1
StrEnum==0.4.15
supabase==1.1.1
supafunc==0.2.3
tinycss2==1.4.0
tinyhtml5==2.0.0
tqdm==4.67.1
truffleHog==2.2.1
truffleHogRegexes==0.0.7
tweepy==4.15.0
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
weasyprint==64.1
webencodings==0.5.1
websockets==12.0
Werkzeug==3.1.3
yarl==1.18.3
zopfli==0.2.3.post1
